Create Task1.1
Task 1.1
For numbers in duodecimal numerical system, the symbols 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, A (ten) and B (eleven) are used. When starting, the application asks the user to input two integers a and b (assume that the user enters integers without errors). Then the application displays all integers in the range from a (inclusive) to b (inclusive), which in their duodecimal representation have exactly two symbols A. Develop a console application that implements the specified functionality.
Note 1. In order to transform string s to int use method int.Parse(s).
Note 2. Output required numbers in decimal (not duodecimal) numerical system.