Create Task1_2.cs
Task  1.2
Let 10-character ISBN is an unique digital code to identify a book, with form: d_1 d_2 d_3 d_4 d_5 d_6 d_7 d_8 d_9 d_10. Digit d_10 is a control one, it is calculated according the condition, that expression 
〖10d〗_1+9d_2+8d_3+...+ 1d_10
(the sum of the products of digits by the weight of their positions) has to be a multiple of 11.
Create an application that prompts the user for a string of 9 digit characters (these are the first nine digits of the ISBN), calculates the check digit, and outputs the resulting ISBN. Do not check the correctness of the user's input - assume that the user does not make errors when entering.
Note 1. Сheck «digit» can be equal to 10. In this case use symbol X to denote it.
Note 2. You can convert any a value to string using a.ToString().