Create Task1_3.cs
Task 1.3
When starting, the application asks the user for the number of elements in the integer array, and then in a loop - the elements themselves (of type int). After entering the elements, the application outputs the original array. Then a new array is formed and output from the original array according to the principle “each value – only once”. For example, for the array [1, 1, 3, 4, 2, 2, 6, 7, 1], the array [1, 3, 4, 2, 6, 7] is formed. Develop a console application that implements the specified functionality.
Note 1. The correctness of the array length may not be controlled.
Note 2. Do not use the standard collections and LINQ!